\name{sumStringNumbers}
\alias{sumStringNumbers}
\title{sumStringNumbers}
\usage{
sumStringNumbers(s, delim = ",")
}
\arguments{
 \item{s}{
  A string of numbers delimeted by ","
 }
 \item{delim}{
  A different dilimeter to be used, default is ","
 }
}
\description{
Prints the sum of a list of numbers.
}
\examples{
sumStringNumbers("1,2")
}
